+++ Plugins/Tweaks/Tweaks/FixScrollLearningBug.hpp
@@ -1,0 +1,18 @@
+#pragma once
+
+#include "API/Types.hpp"
+#include "Common.hpp"
+#include "Services/Hooks/Hooks.hpp"
+
+namespace Tweaks {
+
+class FixScrollLearningBug
+{
+public:
+    FixScrollLearningBug(NWNXLib::Services::HooksProxy* hooker);
+
+private:
+    static BOOL CNWSCreature__LearnScroll_hook(CNWSCreature *, OBJECT_ID);
+};
+
+}